Refactor data packing to web layer.
Move all of the logic that packs and unpacks the data, or rather
translates from the json format to the business objects into the web
layer. This should be the only place where the board size and pieces are
known or configured.

Also clean up calls to board.length or row.length